requirements.txt
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
astroid==2.11.6
asttokens==2.0.5
attrs==21.4.0
autopep8==1.6.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
bs4==0.0.1
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
Django @ file:///tmp/build/80754af9/django_1625585912945/work
django-phonenumber-field==7.0.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
ffmpeg==1.4
Fiona==1.8.21
flake8==4.0.1
fonttools==4.33.3
geopandas==0.11.0
idna==3.3
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isort==5.10.1
jedi==0.17.2
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.8
jsonschema==4.6.1
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.10.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyter-server==1.18.1
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
jupyterlab-widgets==1.1.1
kiwisolver==1.4.3
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
munch==2.5.0
mypy==0.961
mypy-extensions==0.4.3
nbclassic==0.4.2
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nodejs==0.1.1
notebook==6.4.12
notebook-shim==0.1.0
npm==0.1.1
numpy==1.23.0
optional-django==0.1.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
pango==0.7.6
parso==0.7.1
phonenumbers==8.12.55
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2 @ file:///home/builder/ci_310/psycopg2_1640792656405/work
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.12.0
pylint==2.14.4
pyls-mypy @ git+https://github.com/tomv564/pyls-mypy.git@3b105b491a42eec304f19e6ed6da1053d12958c6
pyproj==3.3.1
pyrsistent==0.18.1
python-dateutil==2.8.2
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.4.1
pytoolconfig==1.1.2
pytz==2022.1
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
requests==2.28.1
rope==1.2.0
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
Shapely==1.8.2
six==1.16.0
sklearn==0.0
sniffio==1.2.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
stack-data==0.3.0
style==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
tornado==6.1
traitlets==5.3.0
typing_extensions==4.3.0
ujson==5.4.0
update==0.0.1
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
widgetsnbextension==3.6.1
wrapt==1.14.1
xgboost==1.6.1
yapf==0.32.0
